+++ noodles-bgzf/examples/bgzf_gzi_inspect.rs
@@ -1,14 +1,12 @@
 //! Prints the compressed-uncompressed position pairs of a GZ index.
 
-use std::{env, fs::File, io};
+use std::{env, io};
 
-use noodles_bgzf as bgzf;
+use noodles_bgzf::gzi;
 
 fn main() -> io::Result<()> {
     let src = env::args().nth(1).expect("missing src");
-
-    let mut reader = File::open(src).map(bgzf::gzi::Reader::new)?;
-    let index = reader.read_index()?;
+    let index = gzi::read(src)?;
 
     for record in &index {
         println!("{}\t{}", record.0, record.1);