+++ noodles-bgzf/src/gzi/reader.rs
@@ -44,7 +44,8 @@ where
             usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
         })?;
 
-        let mut offsets = Vec::with_capacity(len);
+        let mut offsets = vec![(0, 0)];
+        offsets.reserve(len);
 
         for _ in 0..len {
             let compressed = self.inner.read_u64::<LittleEndian>()?;
@@ -78,7 +79,10 @@ mod tests {
         ];
 
         let mut reader = Reader::new(&data[..]);
-        assert_eq!(reader.read_index()?, vec![(4668, 21294), (23810, 86529)]);
+        assert_eq!(
+            reader.read_index()?,
+            vec![(0, 0), (4668, 21294), (23810, 86529)]
+        );
 
         Ok(())
     }
@@ -89,7 +93,7 @@ mod tests {
 
         let mut reader = Reader::new(&data[..]);
         let index = reader.read_index()?;
-        assert!(index.is_empty());
+        assert_eq!(index, [(0, 0)]);
 
         Ok(())
     }